+++ src/bibx/sources/base.py
@@ -3,7 +3,7 @@ from typing import Protocol
 from bibx.collection import Collection
 
 
-class CollectionBuilder(Protocol):
+class Source(Protocol):
     """Protocol for classes that build collections of articles."""
 
     def build(self) -> Collection:
