+++ src/bibx/sources/scopus_bib.py
@@ -10,10 +10,10 @@ from bibx.article import Article
 from bibx.collection import Collection
 from bibx.exceptions import MissingCriticalInformationError
 
-from .base import CollectionBuilder
+from .base import Source
 
 
-class ScopusBibCollectionBuilder(CollectionBuilder):
+class ScopusBibSource(Source):
     """Builder for collections of articles from Scopus BibTeX files."""
 
     def __init__(self, *scopus_files: TextIO) -> None:
